+++ init.js
@@ -1,21 +1,29 @@
 /**
- * Copyright 2015 aixigo AG
+ * Copyright 2017 aixigo AG
  * Released under the MIT license.
- * www.laxarjs.org
+ * https://laxarjs.org/license
  */
-require( [
-   'laxar',
-   'laxar-application/var/flows/main/dependencies',
-   'json!laxar-application/var/flows/main/resources.json'
-], function( ax, mainDependencies, mainResources, pageFade ) {
-   'use strict';
 
-   window.laxar.fileListings = {
-      application: mainResources,
-      bower_components: mainResources,
-      includes: mainResources
-   };
+import { create } from 'laxar';
+import artifacts from 'laxar-loader/artifacts?flow=main&theme=cube';
+import debugInfo from 'laxar-loader/debug-info?flow=main&theme=cube';
 
-   ax.bootstrap( mainDependencies );
+import * as angularAdapter from 'laxar-angular-adapter';
 
-} );
+const config = {
+   name: 'finder-demo',
+   router: {
+      query: {
+         enabled: true
+      },
+      navigo: {
+         useHash: true
+      }
+   },
+   theme: 'cube'
+};
+
+create( [ angularAdapter ], artifacts, config )
+   .tooling( debugInfo )
+   .flow( 'main', document.querySelector( '[data-ax-page]' ) )
+   .bootstrap();